# Conflicts:
#	CMakeLists.txt
#	common/Pos2DTemplate.hpp
#	common/Pos3DTemplate.hpp
#	common/enums.hpp
#	common/rng.hpp
#	components/BoundingBox.inl
#	lce
#	loot/tables/stronghold_corridor.hpp
#	structures/gen/stronghold/rolls.cpp
#	structures/gen/village/village.hpp
#	terrain/BalancedSeed.hpp
#	terrain/World.hpp
#	terrain/biomes/biome.hpp
#	terrain/carve/AbstractMapGen.hpp
#	terrain/carve/ChunkGenerator.cpp
#	terrain/carve/ChunkGenerator.hpp

+++ common/worldPicture.cpp
@@ -4,6 +4,7 @@
 
 void WorldPicture::save(const std::string& directory) const {
     const std::string filename = getBiomeImageFileNameFromGenerator(g, directory);
-    stbi_write_png(filename.c_str(), static_cast<int>(myWidth), static_cast<int>(myHeight), myRGBSize, myData,
-                   static_cast<int>(myWidth) * myRGBSize);
+    stbi_write_png(filename.c_str(), static_cast<int>(myWidth), static_cast<int>(myHeight),
+        static_cast<int>(myRGBSize), myData,
+                   static_cast<int>(myWidth * myRGBSize));
 }
